refactor: move theme providers to ui components
