+++ project/components/ui/header.tsx
@@ -6,7 +6,7 @@ import Link from "next/link";
 import { Fragment, useCallback, useEffect, useState } from "react";
 import Button from "../button";
 import LinkButton from "../link-button";
-import { useTheme } from "../theme-provider";
+import { useTheme } from "./theme-provider";
 
 const NAV_LINKS = [
   { href: "/#features", label: "Features" },
